# Use swap operator in radix sort instead of assignment operator