This PR switches an assignment operator to use the swap operator
instead. The swap operator is a simple pointer swap, whereas the
assignment operator is a deep copy, so nearly all of the time
from that assignment is saved, but it is admittedly not the
bottleneck, so only modest performance improvements should be
expected.

Results on 16-node-cs-hdr:
|              | master          | this pr
| ------------ | --------------- | -------
| argsort perf | 11.1145 GiB/sec | 11.9066 GiB/sec

Closes https://github.com/Bears-R-Us/arkouda/issues/1446